Reliably query terminal on high latency connections
Queries to the terminal are sent and the response is read with a timeout
of 100ms. If the latency between the program (running on a remote
server) and the terminal (running on a user's workstation) is larger
than 100ms, this fails and the program may be finished when the response
can finally be read.

This commit replaces the timeout-based approach. Two queries are sent to
the terminal: first, the query we're interested in, but which may not be
supported by the terminal. If it isn't, the terminal will silently
ignore the request and not respond at all.

Next, a cursor position query is sent. This should be supported by all
terminals.

Then a response is read without a timeout. Either the terminal responds
with the cursor position, then we know it does not support the OSC query
and we can abort. If an OSC response is read, we have our result and we
just need to read the cursor position response and discard it.

This approach will work regardless of latency. Since we know that the
terminal will respond (at least to the cursor position request), there's
no need for a timeout, so we can even remove the call to select().

The approach can be tested on Linux by adding e.g. 200ms of artificial
latency via `tc` and running a program on a remote host:

    tc qdisc add dev eth0 root netem delay 200ms 10ms

The latency can be cleared as follows:

    tc qdisc del root dev eth0