Support CLICOLOR_FORCE to force colors without TTY
If termenv is used without a TTY, the output will never be colored. In
some cases, such as in automated tests, there should be a way to force
colored output, even when the environment does not support it.

According to https://bixense.com/clicolors/, CLICOLOR_FORCE variable
does exactly that:

    CLICOLOR_FORCE != 0
	ANSI colors should be enabled no matter what.