Add EnableVirtualTerminalProcessing function
The existing EnableWindowsANSIConsole/RestoreWindowsConsole functions
have a number of limitations:
* They are only defined when termenv is built on Windows, and so require
  the user to use build tags, i.e. multiple source files, to control
  whether or not they are called and are not shown on
  https://pkg.go.dev/github.com/muesli/termenv by default.
* They are hardcoded to set the console mode of stdout, and so fail if
  stdout is not a terminal, e.g. when redirecting the output to a file
  or when run in a Go test.

This commit adds a EnableVirtualTerminalProcessing function with a
different API (to avoid breaking backwards compatibility) that is safe
to call on all platforms, takes an io.Writer as an argument (for output
flexibility). See the comments in the function for more details.