fix: allow to termStatusReport on a PTY
A pty is not in foreground, but it can do the IOCTL ops we need.

My understanding might be wrong, but I think that if the term cannot do
the TCGETS it will fail the same way it would when checking if its in
foreground, so maybe that check is even needed?

refs https://github.com/charmbracelet/wish/issues/262